fix(cosh-ng): [shell] keep provider shell handoffs alive when preexec is lost (#1795)
* fix(cosh-ng): [shell] alias-aware stale-history guard

The bash stale-history guard compared BASH_COMMAND with the raw
history line, so any alias-expanded command (alias ls='ls --color')
mismatched, the preexec marker was swallowed, and the pending shell
handoff deadlocked the Agent run and input routing (issue #1767).

Make the guard alias-aware: _cosh_compact_alias_expanded replays
bash alias expansion over the history original (assignment prefixes,
recursive chains up to depth 10, self-prefix stop, and the
trailing-blank rule where 'alias wrap="env "' keeps expanding the
next word) and the guard accepts a match against either compact form.

The bash/zsh marker scripts move into per-shell owner files
marker/bash.rs and marker/zsh.rs under the registered split plan
(byte-identical extraction; marker.rs stays as a 9-line hub keeping
the shell_host::marker path); the slash registry token test reads
the two per-shell files.


* fix(cosh-ng): [shell] untracked handoff fallback

When a preexec marker is lost with the guard still defeated (deep
alias chains, unknown shells), an emitted shell handoff stayed
pending forever and blocked the Agent run and input routing.

Add a prompt-boundary fallback: ShellHandoffState records the event
index at emission, and close_untracked_shell_handoffs closes the
front pending handoff once the first relevant boundary after that
index is a ShellReady - a CommandStarted seen first defers to the
tracked block path, and events after an already-reached boundary
never veto it (the event list is cumulative). Closure emits degraded
evidence (completed_untracked, exit code -1, no output), marks the
provider tool completed, and feeds the existing continuation
recovery; the provider notice states the result must not be treated
as a command outcome.

Organization: the handoff row builders (tracked and untracked) live
in the new activity/shell_handoff.rs owner file with runtime.rs as a
facade; the cross-owner status token SHELL_HANDOFF_UNTRACKED_STATUS
and the ShellHandoffRequest contract move to types/shell_handoff.rs;
the ui status label matches the shared constant; i18n gains the
untracked status label in en/zh.


* test(cosh-ng): [shell] pin handoff guard contracts

Unit tests pin the untracked fallback semantics: closure on the
first ShellReady after emission, deferral when a CommandStarted
comes first, survival of commands that start after the boundary
was already reached, idempotent replay of cumulative event lists,
shared prompt boundary for same-cycle handoffs, pending state for
handoffs emitted past the boundary, and untouched interactive
handoffs; approval tests cover the emission index bookkeeping.

PTY integration tests pin the alias-aware guard: single alias,
recursive chain, assignment prefix, compound, pipeline, quoted
argument, trailing-blank chain (alias wrap='env '; wrap ll <dir>),
and a polluted-IFS session - every command stays tracked with its
history-original text and exit 0. A raw CLI end-to-end case drops
the preexec marker on purpose and asserts the provider turn still
completes on degraded untracked evidence.


* test(cosh-ng): [shell] raise cosh-shell test floor

This PR adds source tests for the handoff guard and fallback, so
the cosh-shell source floor moves 2469 -> 2528 on the current main
baseline. The lib/bin overlap ceiling is untouched: none of the new
tests appear in the intersection.


* docs(cosh-ng): [shell] sync organization ledger

large-file-inventory.md: drop the shell_host/marker.rs entry (split
into per-shell owner files below the threshold) and refresh
activity/runtime.rs (769 after extracting activity/shell_handoff.rs).

decisions.md (new in-repo registry): D13 registers the pre-existing
activity -> runtime reverse dependency as a temporary exception with
removal conditions; D14 registers SHELL_HANDOFF_UNTRACKED_STATUS as
cross-owner string-contract debt following the D11 precedent.


---------
