+++ src/cosh-ng/crates/cosh-shell/src/activity/mod.rs
@@ -1,6 +1,7 @@
 pub(super) mod runtime;
 mod runtime_output;
 mod runtime_render;
+mod shell_handoff;
 mod tool_invocation;
 mod tool_result_summary;
