+++ src/cosh-ng/crates/cosh-shell/tests/raw_cli/host_executed.rs
@@ -1,6 +1,181 @@
 use super::*;
 
 #[test]
+fn raw_cli_host_executed_pipeline_alias_handoff_never_deadlocks() {
+    // R6 contract (specs/shell-handoff-preexec-loss, validation A15): an
+    // aliased word in a non-leading pipeline element may still lose the
+    // preexec marker (DEBUG trap firing order is unspecified). Whichever
+    // path is taken — tracked normally or closed by the untracked
+    // fallback — the handoff must reach the provider and the turn must
+    // finish instead of deadlocking.
+    let home = temp_shell_home("qwen-pipeline-alias-handoff");
+    let bin_dir = home.join("bin");
+    fs::create_dir_all(&bin_dir).unwrap();
+    let co_path = bin_dir.join("co");
+    write_executable(
+        &co_path,
+        r#"#!/bin/sh
+read -r init
+printf '%s\n' '{"type":"control_response","response":{"subtype":"success","request_id":"init-1","response":{"subtype":"initialize","capabilities":{"can_handle_can_use_tool":true,"can_handle_host_executed_shell_tool_result":true}}}}'
+printf '%s\n' '{"type":"system","subtype":"init","session_id":"sess-pipeline","model":"qwen-test"}'
+read -r user_message
+case "$user_message" in
+  *provider-pipeline-alias*)
+    printf '%s\n' '{"type":"control_request","request_id":"ctrl-1","request":{"subtype":"can_use_tool","tool_name":"run_shell_command","input":{"command":"echo needle | grep needle"},"tool_use_id":"toolu-1"}}'
+    if IFS= read -r response; then
+      case "$response" in
+        *'"behavior":"host_executed_shell"'*)
+          printf '%s\n' '{"type":"assistant","session_id":"sess-pipeline","message":{"content":[{"type":"text","text":"Pipeline handoff result received."}]}}'
+          printf '%s\n' '{"type":"result","subtype":"success","session_id":"sess-pipeline","is_error":false,"result":"done"}'
+          exit 0
+          ;;
+      esac
+    fi
+    printf '%s\n' '{"type":"result","subtype":"error","session_id":"sess-pipeline","is_error":true,"result":"missing host_executed_shell result"}'
+    exit 1
+    ;;
+esac
+printf '%s\n' '{"type":"result","subtype":"success","session_id":"sess-pipeline","is_error":false,"result":"ignored"}'
+"#,
+    );
+    let old_path = std::env::var("PATH").unwrap_or_default();
+    let path = format!("{}:{old_path}", bin_dir.display());
+    let home_str = home.to_string_lossy().to_string();
+    let output = run_raw_cli_serial_with_args_env_and_delayed_input(
+        "qwen",
+        &[],
+        &[
+            ("HOME", &home_str),
+            ("PATH", &path),
+            ("COSH_SHELL_DEBUG", "1"),
+        ],
+        vec![
+            (b"/mode approval trust confirm\n".to_vec(), Duration::ZERO),
+            (
+                b"alias grep='grep --color=auto'\n".to_vec(),
+                Duration::from_millis(1_500),
+            ),
+            (
+                b"?? provider-pipeline-alias\n".to_vec(),
+                Duration::from_millis(1_500),
+            ),
+            (
+                b"/details handoff-1\n".to_vec(),
+                Duration::from_millis(6_000),
+            ),
+            (b"exit\n".to_vec(), Duration::from_millis(500)),
+        ],
+    );
+    let _ = fs::remove_dir_all(&home);
+
+    assert!(output.contains("Bash tool sent to shell"), "{output}");
+    // The command really ran.
+    assert!(output.contains("needle"), "{output}");
+    // Whichever closure path was taken, the handoff must be closed with
+    // evidence and the provider turn must finish (no deadlock).
+    assert!(
+        output.contains("status: completed") || output.contains("status: completed_untracked"),
+        "{output}"
+    );
+    assert!(
+        output.contains("Pipeline handoff result received."),
+        "{output}"
+    );
+    assert!(
+        !output.contains("missing host_executed_shell result"),
+        "{output}"
+    );
+    assert!(!output.contains("Agent timed out:"), "{output}");
+}
+
+#[test]
+fn raw_cli_host_executed_untracked_handoff_closes_and_recovers() {
+    // Simulates a real lost-preexec condition deterministically: the user
+    // (or a rogue rc snippet) clears the DEBUG trap, so preexec is never
+    // emitted for the injected handoff while PROMPT_COMMAND still emits
+    // precmd (ShellReady). The prompt boundary must close the pending
+    // handoff with degraded untracked evidence instead of deadlocking the
+    // Agent (specs/shell-handoff-preexec-loss D2, validation A9/A10).
+    let home = temp_shell_home("qwen-untracked-handoff");
+    let bin_dir = home.join("bin");
+    fs::create_dir_all(&bin_dir).unwrap();
+    let co_path = bin_dir.join("co");
+    write_executable(
+        &co_path,
+        r#"#!/bin/sh
+read -r init
+printf '%s\n' '{"type":"control_response","response":{"subtype":"success","request_id":"init-1","response":{"subtype":"initialize","capabilities":{"can_handle_can_use_tool":true,"can_handle_host_executed_shell_tool_result":true}}}}'
+printf '%s\n' '{"type":"system","subtype":"init","session_id":"sess-untracked","model":"qwen-test"}'
+read -r user_message
+case "$user_message" in
+  *provider-untracked-handoff*)
+    printf '%s\n' '{"type":"control_request","request_id":"ctrl-1","request":{"subtype":"can_use_tool","tool_name":"run_shell_command","input":{"command":"df -h"},"tool_use_id":"toolu-1"}}'
+    if IFS= read -r response; then
+      case "$response" in
+        *'"behavior":"host_executed_shell"'*'execution was not tracked'*)
+          printf '%s\n' '{"type":"assistant","session_id":"sess-untracked","message":{"content":[{"type":"text","text":"Untracked handoff result received."}]}}'
+          printf '%s\n' '{"type":"result","subtype":"success","session_id":"sess-untracked","is_error":false,"result":"done"}'
+          exit 0
+          ;;
+      esac
+    fi
+    printf '%s\n' '{"type":"result","subtype":"error","session_id":"sess-untracked","is_error":true,"result":"missing untracked declaration"}'
+    exit 1
+    ;;
+esac
+printf '%s\n' '{"type":"result","subtype":"success","session_id":"sess-untracked","is_error":false,"result":"ignored"}'
+"#,
+    );
+    let old_path = std::env::var("PATH").unwrap_or_default();
+    let path = format!("{}:{old_path}", bin_dir.display());
+    let home_str = home.to_string_lossy().to_string();
+    let output = run_raw_cli_serial_with_args_env_and_delayed_input(
+        "qwen",
+        &[],
+        &[
+            ("HOME", &home_str),
+            ("PATH", &path),
+            ("COSH_SHELL_DEBUG", "1"),
+        ],
+        vec![
+            (b"/mode approval auto\n".to_vec(), Duration::ZERO),
+            (b"trap - DEBUG\n".to_vec(), Duration::from_millis(800)),
+            (
+                b"?? provider-untracked-handoff\n".to_vec(),
+                Duration::from_millis(800),
+            ),
+            (
+                b"/details handoff-1\n".to_vec(),
+                Duration::from_millis(6_000),
+            ),
+            (b"exit\n".to_vec(), Duration::from_millis(500)),
+        ],
+    );
+    let _ = fs::remove_dir_all(&home);
+
+    assert!(output.contains("Auto-approved req-1"), "{output}");
+    assert!(output.contains("Bash tool sent to shell"), "{output}");
+    // The command really ran even though tracking was lost.
+    assert!(output.contains("Filesystem"), "{output}");
+    // Degraded closure instead of a stuck handoff (A9).
+    assert!(output.contains("status: completed_untracked"), "{output}");
+    assert!(output.contains("preexec_marker_missing"), "{output}");
+    assert!(output.contains("exit_code: unavailable"), "{output}");
+    assert!(output.contains("output_id: <none>"), "{output}");
+    // Provider received the untracked declaration and finished the turn (A10:
+    // the run resumes instead of waiting on the handoff forever).
+    assert!(
+        output.contains("Untracked handoff result received."),
+        "{output}"
+    );
+    assert!(
+        !output.contains("missing untracked declaration"),
+        "{output}"
+    );
+    assert!(!output.contains("Agent timed out:"), "{output}");
+}
+
+#[test]
 fn raw_cli_host_executed_shell_result_continues_same_provider_turn() {
     let home = temp_shell_home("qwen-host-executed-shell");
     let bin_dir = home.join("bin");
